+++ RELEASENOTES.md
@@ -3,6 +3,7 @@
 ## 0.0.2
 
 * Fixed directory traversal bug.
+* Added optimization for "contents contains" queries.
 
 ## 0.0.1